Rendering both nodes and edges from one component.
